DL: Improving [git] handling / logging
With Guen's suggestion that we have multiple repos and add/commit
as well as capture the version hash, it has become
necessary to expand the git logging to beyond the current working
directory.  To that end, I've made GitUtils.m which contains a
class that:
1) getting old status
2) adding ALL new files (except the ones in .gitignore)
3) committing with given message
4) getting new status
5) get the version hash (unique string for each commit)

With this, we should be able to snapshot any arbitrary git repo.

I still need to test git on accessing different directories than
just '.' current working directory.